requirements.txt
# Phoenix Agent v2 Requirements
# Core dependencies for document processing and PDF generation

# PDF processing
PyMuPDF>=1.23.0
reportlab>=4.0.0

# Image processing and computer vision
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0

# Machine Learning (for YOLO and Nougat integration)
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0

# OCR and document understanding
nougat-ocr>=0.1.0
transformers>=4.30.0

# Google AI (for Gemini integration)
google-generativeai>=0.3.0

# Utilities
pathlib2>=2.3.7
typing-extensions>=4.5.0
argparse>=1.4.0

# Optional: For enhanced PDF manipulation
PyPDF2>=3.0.0

# Optional: For better image processing
scikit-image>=0.21.0

# Optional: For advanced text processing
nltk>=3.8.1
spacy>=3.6.0 